+++ src/Request/Init.php
@@ -39,7 +39,7 @@ class Init
      * Valid values (see also `$validServices`):
      *  - assess
      *  - author
-     *  - author-aide
+     *  - authoraide
      *  - data
      *  - events
      *  - items
@@ -110,7 +110,7 @@ class Init
      * Service names that are valid for `$service`
      * @var array
      */
-    private $validServices = ['assess', 'author', 'author-aide', 'data', 'events', 'items', 'questions', 'reports'];
+    private $validServices = ['assess', 'author', 'authoraide', 'data', 'events', 'items', 'questions', 'reports'];
 
     /**
      * @var SignatureFactory
@@ -247,7 +247,7 @@ class Init
                     $this->requestPacket;
                 break;
             case 'author':
-            case 'author-aide':
+            case 'authoraide':
             case 'items':
             case 'reports':
                 // Add the security packet (with signature) to the output